[PATCH] USB: fix functions to match prototypes

+++ drivers/usb/media/stv680.c
@@ -1313,7 +1313,7 @@ static int stv680_mmap (struct file *file, struct vm_area_struct *vma)
 	return 0;
 }
 
-static int stv680_read (struct file *file, char *buf,
+static ssize_t stv680_read (struct file *file, char *buf,
 			size_t count, loff_t *ppos)
 {
 	struct video_device *dev = file->private_data;
